fixed test with abstraction refactor
